+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-* Update autoscaling.
+* Update toc file.