+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-* Added launch button to DataBroker.
+* Update to 1.13.6